+++ source/CHGlobals.h
@@ -1,0 +1,11 @@
+//
+//  DMGlobals.h
+//  Dimensions
+//
+//  Created by Zach Waugh on 10/16/10.
+//  Copyright 2010 zachwaugh.com. All rights reserved.
+//
+
+extern NSString * const DMCopyFormatKey;
+extern NSString * const DMPrimaryOverlayColorKey;
+extern NSString * const DMAlternateOverlayColorKey;