+++ source/CHGlobals.m
@@ -1,0 +1,14 @@
+//
+//  DMGlobals.m
+//  Dimensions
+//
+//  Created by Zach Waugh on 10/16/10.
+//  Copyright 2010 zachwaugh.com. All rights reserved.
+//
+
+#import "CHGlobals.h"
+
+
+NSString * const DMCopyFormatKey = @"copyFormat";
+NSString * const DMPrimaryOverlayColorKey = @"primaryOverlayColor";
+NSString * const DMAlternateOverlayColorKey = @"alternateOverlayColor";