+++ source/CHPreferences.h
@@ -1,0 +1,19 @@
+//
+//  DMPreferences.h
+//  Dimensions
+//
+//  Created by Zach Waugh on 10/16/10.
+//  Copyright 2010 zachwaugh.com. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+
+@interface CHPreferences : NSObject
+{
+
+}
+
++ (void)registerDefaults;
+
+@end