+++ resources/CHANGELOG.md
@@ -1,8 +1,10 @@
-0.9.2b
-- Changed name from Crosshairs to Dimensions
-- Delete key clears overlay
+0.9.5b
+- Delete key now clears overlay
 - Spacebar takes screenshot
 - Fixed text shadow color on white overlay
+- Fixed issue where arrow cursor would show instead of crosshairs
+- Hides handles on resize
+- Fixed some resizing bugs
 
 0.9.1b
 - Changed "Activate" to "Show" in dropdown menu